PAY-2115: Code Refactor & cleanup unnecessary
