--- src/main/java/uk/gov/hmcts/reform/bulkscanning/service/BulkScanConsumerService.java
@@ -1,11 +1,0 @@
-package uk.gov.hmcts.reform.bulkscanning.service;
-
-import uk.gov.hmcts.reform.bulkscanning.model.entity.Envelope;
-import uk.gov.hmcts.reform.bulkscanning.model.request.CaseReferenceRequest;
-import uk.gov.hmcts.reform.bulkscanning.model.request.BulkScanPaymentRequest;
-
-public interface BulkScanConsumerService {
-    Envelope saveInitialMetadataFromBs(BulkScanPaymentRequest bsPaymentRequest);
-    String updateCaseReferenceForExceptionRecord (String exceptionRecordReference, CaseReferenceRequest caseReferenceRequest);
-    String markPaymentAsProcessed(String dcn);
-}
